fix(git init): set uptstream if not set
The upstream was set for pushing but nor for pulling. Also, it was not
the best way to set the upstream. Here, we are checking the default
remote branch and we use it to define upstream. So, even if the remote
has several branch, we can get the branch to track.