fix(add dotfiles): check if file using home shorthand exists
This bug was introduced by #066c7c6.
From CLI everything works as expected. However, from the menu, if a
path contains `~/` it is somehow escaped inside the test command so the
file is not recognized.
I tried to remove the prefix and suffix (`'`) before testing the file
existence but it is not working...
So I ended up by moving the test inside the function to obtain absolute
path even if I repeat myself...