chore(add): copy the dotfiles based on XDG paths and allow other files
Instead of copying the path of the current operating system, I think
it will be more portable if Dotig is based on the XDG Specification.
So, when a dotfile is add, the `XDG_xxx_HOME` part is checked to place
the backup file in a corresponding folder.
If the file is not in a XDG path but still in `$HOME`, it will be
copied by mirroring the path on the current operating system.
Then symlinks will be created between the backup and the original
location.

Some people also backup some root files. So, Dotig now accept these
files. The original path will be append to the root repo. Example:
`/home/username/.dotfiles/etc/nanorc`
However the symlinks will not be created in this case. I think
it is not safe to replace the root files in an automated way.