chore: improve help menu
I add the possibility to pass verbose option before commands.
It was not clear that DOTFILES is an environment variable, so I reword
the help menu to improve the description.
I also add some # before titles to improve distinction with content. It
is useful for "CONFIGURATION" and "DOTFILES" since both are capitalized