chore: remove verbose condition when removing broken symlinks
By using the menu, the user can understand that no broken symlinks was
found since a prompt to return to the menu appears. However, from CLI,
it is not clear if the command was successful or if an error occurred.
So, I remove the verbose condition. Now, if there are no broken
symlinks, Dotig will always print a message.