refactor: add a function to handle verbose msg
The various "if is verbose" weigh down the code. By using a custom
function to handle verbose messages, I think that improves readability.
Even when success or other messages handled with a function are used, I
think it remains human readable.
It also prevents a misuse of shorthand "if" with "set -e". "is_verbose"
with the "if" shortened form can result to an exit according to its
place in the code.