feat: add an interactive mode
Instead of asking each time if the repo is correct, I implemented an
interactive option. Without it, Dotig assumes the repo is set. If not,
an error is displayed. With the interactive option ( `--interactive` or
`-i`), user will be invited to configure its repo.
This option can also print additional dialogs such as asking user if
he wants to remove broken symlinks during an update (even if the `-cb`
or `--check-broken` option is not set).