refactor(status)!: replace no-status option with status option
Instead of printing the status by default, I think it is preferable to
display it only through an option. It allows Dotig to launch quickly
and the status is not always necessary.
It also makes it possible to keep the names of shortened options
consistent (only one letter).

BREAKING CHANGE: `dotig -ns` and `dotig --no-status` replaced by
`dotig -s` and `dotig --status`. The status is no longer displayed by
default before the Dotig menu.