refactor(remove)!: rename broken symlinks option
Replace `--with-broken` (and `-wb`) by `--broken-symlinks` (and `-b`).
I think the name is more explicit. Also, it allows me to keep the
options consistent: all shortened options use only one letter. And the
same name is used for the `update` command, so the user will not be
confused.

BREAKING CHANGE: `dotig rm --with-broken` and `dotig rm -wb` are
replaced by `dotig rm --broken-symlinks` and `dotig rm -b`.